moved external links tests to separate file
