+++ lmfdb/tests/test_homepage.py
@@ -62,17 +62,3 @@ class HomePageTest(LmfdbTest):
         self.check(homepage, "/universe", 'universe')
         # removed in PR #1167
         #self.check(homepage, "/knowledge/", 'Recently modified Knowls')
-
-    # Box 6
-    def test_box6(self):
-        r"""
-        Check that the links in Box 6 work.
-        """
-        homepage = self.tc.get("/").get_data(as_text=True)
-        self.check_external(homepage, "https://github.com/LMFDB/lmfdb", 'Modular Forms Database')
-        # I could not get this one to work - AVS
-        #self.check_external(homepage, "http://www.sagemath.org/", 'mathematics software system')
-        self.check_external(homepage, "http://pari.math.u-bordeaux.fr/", 'PARI/GP is a widely used computer algebra system')
-        # I could not get this one to work -- JEC
-        #self.check_external(homepage, "http://magma.maths.usyd.edu.au/magma/", 'Magma is a large, well-supported software package')
-        self.check_external(homepage, "https://www.python.org/", 'Python Logo')
